+++ examples/full_example/lib/widgetbook.dart
@@ -23,7 +23,10 @@ class WidgetbookApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return Widgetbook.material(
       addons: [
+        DeviceFrameAddon(devices: Devices.ios.all),
+        GridAddon(100),
         AlignmentAddon(),
+        ZoomAddon(),
       ],
       directories: [
         WidgetbookFolder(