+++ packages/widgetbook_cli/bin/cli_production.dart
@@ -1,7 +1,7 @@
-import 'app/main.dart' as cli;
-import 'flavor/flavor.dart';
+import 'package:widgetbook_cli/widgetbook_cli.dart';
+
+import 'main.dart' as cli;
 
 void main(List<String> args) {
-  Flavor().strategy = DeploymentStrategy.production;
-  cli.main(args);
+  cli.main(args, ProductionEnv());
 }